Migrate UCCSD-HF unittests
On a side-note: do we really need these extensive UCCSD and HF test
cases?
